Generalize shader tests
